.vercel - Copia/cache/index_py/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.1
email-validator==2.1.0.post1
typing-extensions==4.9.0
anyio==4.2.0
